refactor(playback): extract chapter queue creation
